refactor(connection-string): Rename `pg-connection-string` package to `gaussdb--connection-string`
