+++ packages/gaussdb-connection-string/.coveralls.yml
