+++ packages/gaussdb-connection-string/.gitignore
