+++ packages/gaussdb-connection-string/.mocharc.json
