+++ packages/gaussdb-connection-string/LICENSE
