+++ packages/gaussdb-connection-string/index.d.ts
