+++ packages/gaussdb-connection-string/test/clientConfig.ts
