+++ packages/gaussdb-connection-string/test/example.ca
