+++ packages/gaussdb-connection-string/test/example.cert
