+++ packages/gaussdb-connection-string/test/parse.ts
